IAT-3078: Refactoring due to changes in common
