+++ gradle.properties
@@ -12,7 +12,7 @@ group=org.opentestsystem.ap
 
 version=0.2.196-SNAPSHOT
 
-commonVersion=0.4.214
+commonVersion=0.4.215-SNAPSHOT
 
 configServerVersion=0.0.1
