+++ src/main/java/org/opentestsystem/ap/irs/service/RenderingService.java
@@ -16,10 +16,10 @@
 package org.opentestsystem.ap.irs.service;
 
 import lombok.extern.slf4j.Slf4j;
-import org.opentestsystem.ap.common.preview.PreviewManager;
 import org.opentestsystem.ap.common.model.PreviewData;
 import org.opentestsystem.ap.common.model.render.RenderResponse;
-import org.opentestsystem.ap.common.util.SecurityUtil;
+import org.opentestsystem.ap.common.preview.PreviewManager;
+import org.opentestsystem.ap.common.security.SecurityUtil;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
