+++ src/test/java/org/opentestsystem/ap/irs/service/RenderingServiceIT.java
@@ -4,9 +4,9 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
-import org.opentestsystem.ap.common.preview.PreviewManager;
 import org.opentestsystem.ap.common.model.render.RenderResponse;
-import org.opentestsystem.ap.common.util.SecurityUtil;
+import org.opentestsystem.ap.common.preview.PreviewManager;
+import org.opentestsystem.ap.common.security.SecurityUtil;
 import org.opentestsystem.ap.irs.util.IRSTestUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.TestConfiguration;
