+++ src/test/java/org/opentestsystem/ap/irs/util/IRSTestUtil.java
@@ -2,9 +2,9 @@ package org.opentestsystem.ap.irs.util;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.opentestsystem.ap.common.model.ItemBankUser;
 import org.opentestsystem.ap.common.model.ItemFactory;
 import org.opentestsystem.ap.common.model.render.RenderResponse;
+import org.opentestsystem.ap.common.security.model.ItemBankUser;
 
 /**
  * Created by alexponce on 7/3/17.
